Rename to pedpy
