+++ README.md
@@ -1,6 +1,6 @@
-# Pedestrian-Trajectory-Analyzer (PiTA)
+# pedpy
 
-`PiTA` is a python library to analyze pedestrian trajectories. 
+`pedpy` is a python module for pedestrian movement analysis. 
 
 This repo is a port from the original `JPSreport` to a Python implementation, and will provide the same functionalities.
 
@@ -9,7 +9,7 @@ This repo is a port from the original `JPSreport` to a Python implementation, an
 For setting up your Python Environment a Python version >= 3.8 is recommended (our code is tested with 3.8).
 To avoid conflicts with other libraries/applications the usage of virtual environments is recommended, see [Python Documentation](https://docs.python.org/3/library/venv.html) for more detail.
 
-Now the needed libraries for `PiTA` need to be installed.
+Now the needed libraries for `pedpy` need to be installed.
 You can use the provided `requirements.txt` to do so:
 
 ```bash
