+++ pedpy/__init__.py
@@ -1,7 +1,6 @@
 from . import _version
 from .data.geometry import Geometry
 from .data.trajectory_data import TrajectoryData, TrajectoryUnit
-from .io.geometry_parser import parse_geometry
 from .io.trajectory_loader import load_trajectory
 from .methods.density_calculator import (
     compute_classic_density,
