+++ setup.cfg
@@ -1,7 +1,7 @@
 [versioneer]
 VCS = git
 style = pep440-pre
-versionfile_source = analyzer/_version.py
-versionfile_build = analyzer/_version.py
+versionfile_source = pedpy/_version.py
+versionfile_build = pedpy/_version.py
 tag_prefix =
 parentdir_prefix =
